requirements.txt
requests>=2.31.0
jinja2>=3.1.2
markdown>=3.5.1
reportlab>=4.0.7
pytest>=7.4.3
loguru>=0.7.2
pyyaml>=6.0.1
yt-dlp>=2023.12.30
youtube-transcript-api>=0.6.1
openai>=1.12.0
aiohttp>=3.9.1
prefect>=2.14.0
pillow>=10.2.0
ultralytics>=8.1.0  # for YOLO
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0
python-dotenv>=1.0.0
fastapi>=0.109.0  # for REST API endpoints
uvicorn>=0.27.0  # for running FastAPI server 